# Conflicts:
#	xzxms-purchase/purchase-controller/src/main/java/com/xzx/xzxms/purchase/controller/PurchasePlanController.java

+++ xzxms-purchase/purchase-service/src/main/java/com/xzx/xzxms/purchase/service/PurchasePlanService.java
@@ -38,7 +38,7 @@ public interface PurchasePlanService {
      * @param purchaseItems
      * @param itemNum
      */
-    String insertItemService(PurchaseItems purchaseItems, int itemNum);
+    String insertItemService(PurchaseItems purchaseItems, Double itemNum);
 
     /**
      * 根据购买项id查找其供应商
